POST https://stem-kintai.vercel.app/kiosk 500 (Internal Server Error) M
